create a node to uncompress the compressed camera feed on the workstation side
The node subscribes to the /camera/color/image_raw/compressed topic and use CvBridge to decompressw the image into an image message (sensor_msgs.msgs.Image) and publish on the /camera/color/uncompressed topic. I noticed that the uncompressed topic should not be namespaced as /camera/color/image_raw/uncompressed because RViz was trying to look for the camera info under /camera/color/image_raw/camera_info.

Modified the package.xml and setup.py file for the node to be recognized as an executable